The great backend overhaul.
The module Fut lives in its own fut ocamlfind package and is now
Unix-free (link-wise, as we still need the cmi during compilation
because the backend signature mentions Unix.file_descr). To use Fut
the fut package should be used along with a backend package,
fut.select or fut.jsoo for now.

Each backend is confined to its own cm[x]a and has its own ocamlfind
package. Use the backend package with fut to use the backend. Unix
futures are moved to their own Futu module which is also a seperate
package and is implemented over the public interface of Fut.

The jsoo backend is currently stubbed but the base combinators tests,
which don't need the runtime system, pass in the browser.