Get the backend timer interface right.
Absolute times have not their place here, if you want to schedule a
future at an absolute time it should be able to survive a crash so you
need to write these deadlines somewhere on disk and reload them when
you restart a program. This functionality can be derived and should
not be part of Fut's core but in a third party library. Moreover
backends should use a monotonic clock and not gettimeofday(2) like the
select(2) backend does now as gettimeofday(2) can go back in time !

We also try to get timer abortion right in preparation for
the new abort semantics.