Revert commit e97fb3d4eabd9df.
I don't think raising Invalid_argument if the future's promise
is already set is very practical for the clients given that
the future can be aborted behind the scene. The client can
get a notification through the ~abort argument of the promise
when that's the case and avoid trying to set the promise but
sometimes it may be more practical to just let it run and
try to set the promise anyways (even though it will have
no effect.