First implementation of the new abort semantics.
As described in fut.ml there are two ways for a future to abort when
it is set to never determine.

1) When its dependent futures behave such that its semantics says it
should be set to never determine. In that case we just call the
future's abort function, this is a *shallow abort* and is mainly used
by applicative combinators (if they are not deeply aborted).

2) When the client explicitely calls Fut.abort on an undetermined
future or that it uses effectful combinators like Fut.pick that abort
one of their arguments. In that case we call the future's abort
function and walk up the whole dependency chain of the future and
abort each dependent. This violence can be mitigated with Fut.protect
that cuts the dependency chain.

So the choice is to make futures deeply abortable by default.

The current implementation needs refactoring for undetermined futures,
the [stop_wait] functional field to remove waiters and [deep_aborts]
that keep the dependencies should be merged toghether in the GADT in
the same case (and with better names) so that we have the waiter and
the dep together (rather than the waiter in the [stop_wait] as we have
now). In the implementation of the combinators we should always link
waiters and their dep like [waits] does now.