Clean implementation of deep aborts.
Everything is now factored out in the notion of dependency sets for a
future in which we hold the futures we wait on tupled with the waiter
we registered in it. This allows to stop waiting on the future using a
data structure rather than a closure (the [stop_wait] field we had,
thanks GADTs) and we have the dependency at hand for deep aborts.

That notion of dependency is the sole thing that needs to be used by
the combinators. Except for the special case Fut.protect, combinators
do not need to muck around with the undet's fields anymore, resulting
in much more understandable an trivial code.

Some work is still needed around waiter execution which is not T.R. a
blowing stack test case should be made.  Waiter execution should
either go through a queue to avoid the depth first, direct, execution
that we have now. Another possibility would be to thread a
continuation in the waiters, that could actually be useful for also
guaranteeing that deep aborts won't stack overflow.